+++ src/main/java/dev/dubhe/anvilcraft/advancements/criterion/DevourerDevourTrigger.java
@@ -1,4 +1,4 @@
-package dev.dubhe.anvilcraft.advancements.criteron;
+package dev.dubhe.anvilcraft.advancements.criterion;
 
 import com.mojang.serialization.Codec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
@@ -12,7 +12,7 @@ import net.minecraft.world.level.block.Block;
 
 import java.util.Optional;
 
-public class PlacerPlaceTrigger extends SimpleCriterionTrigger<PlacerPlaceTrigger.TriggerInstance> {
+public class DevourerDevourTrigger extends SimpleCriterionTrigger<DevourerDevourTrigger.TriggerInstance> {
     @Override
     public Codec<TriggerInstance> codec() {
         return TriggerInstance.CODEC;
@@ -22,18 +22,18 @@ public class PlacerPlaceTrigger extends SimpleCriterionTrigger<PlacerPlaceTrigge
         this.trigger(player, (instance) -> instance.matches(block));
     }
 
-    public record TriggerInstance(Optional<ContextAwarePredicate> player, Optional<BlockPredicate> block) implements SimpleCriterionTrigger.SimpleInstance {
+    public record TriggerInstance(Optional<ContextAwarePredicate> player, Optional<BlockPredicate> block) implements SimpleInstance {
         public static final Codec<TriggerInstance> CODEC = RecordCodecBuilder.create((instance) -> instance.group(
             EntityPredicate.ADVANCEMENT_CODEC.optionalFieldOf("player").forGetter(TriggerInstance::player),
             BlockPredicate.CODEC.optionalFieldOf("block").forGetter(TriggerInstance::block)
         ).apply(instance, TriggerInstance::new));
 
-        public static Criterion<TriggerInstance> placeBlock(Block block) {
-            return placeBlock(BlockPredicate.Builder.block().of(block));
+        public static Criterion<TriggerInstance> devourBlock(Block block) {
+            return devourBlock(BlockPredicate.Builder.block().of(block));
         }
 
-        public static Criterion<TriggerInstance> placeBlock(BlockPredicate.Builder block) {
-            return ModCriterionTriggers.PLACER_PLACE_BLOCK.get().createCriterion(new TriggerInstance(Optional.empty(), Optional.of(block.build())));
+        public static Criterion<TriggerInstance> devourBlock(BlockPredicate.Builder block) {
+            return ModCriterionTriggers.DEVOURER_DEVOUR_BLOCK.get().createCriterion(new TriggerInstance(Optional.empty(), Optional.of(block.build())));
         }
 
         public boolean matches(Block block) {
